+++ src/Inline/Element/Image.php
@@ -14,50 +14,27 @@
 
 namespace League\CommonMark\Inline\Element;
 
+use League\CommonMark\Util\ArrayCollection;
+
 class Image extends AbstractWebResource
 {
     /**
-     * @var InlineCollection
-     */
-    protected $altText;
-
-    /**
      * @param string $url
-     * @param InlineCollection|string $label
+     * @param ArrayCollection|string $label
      * @param string $title
      */
     public function __construct($url, $label = '', $title = '')
     {
         parent::__construct($url);
 
-        $this->setAltText($label);
+        if (is_string($label)) {
+            $this->inlineContents = new ArrayCollection(array(new Text($label)));
+        } else {
+            $this->inlineContents = $label;
+        }
 
         if (!empty($title)) {
             $this->data['title'] = $title;
         }
-    }
-
-    /**
-     * @return InlineCollection
-     */
-    public function getAltText()
-    {
-        return $this->altText;
-    }
-
-    /**
-     * @param InlineCollection|string $label
-     *
-     * @return $this
-     */
-    public function setAltText($label)
-    {
-        if (is_string($label)) {
-            $this->altText = new InlineCollection(array(new Text($label)));
-        } else {
-            $this->altText = $label;
-        }
-
-        return $this;
     }
 }